+++ src/app.jsx
@@ -1,14 +1,14 @@
-import jwtDecode from 'jwt-decode';
-import React, { useEffect, useRef, useState } from 'react';
-import Screen from './screen';
-import ContentService from './service/contentService';
-import ConfigLoader from './config-loader';
-import ReleaseLoader from './release-loader';
-import Logger from './logger/logger';
-import './app.scss';
-import localStorageKeys from './local-storage-keys';
-import fallback from './assets/fallback.png';
-import idFromPath from './id-from-path';
+import jwtDecode from "jwt-decode";
+import React, { useEffect, useRef, useState } from "react";
+import Screen from "./screen";
+import ContentService from "./service/contentService";
+import ConfigLoader from "./config-loader";
+import ReleaseLoader from "./release-loader";
+import logger from "./logger/logger";
+import localStorageKeys from "./local-storage-keys";
+import fallback from "./assets/fallback.png";
+import idFromPath from "./id-from-path";
+import "./app.scss";
 
 /**
  * App component.
@@ -26,7 +26,7 @@ function App() {
   const releaseTimestampIntervalTimeoutDefault = 1000 * 60 * 10;
 
   const [running, setRunning] = useState(false);
-  const [screen, setScreen] = useState('');
+  const [screen, setScreen] = useState("");
   const [bindKey, setBindKey] = useState(null);
   const [refreshingToken, setRefreshingToken] = useState(false);
   const timeoutRef = useRef(null);
@@ -46,7 +46,7 @@ function App() {
   const appStyle = {};
 
   if (!debug) {
-    appStyle.cursor = 'none';
+    appStyle.cursor = "none";
   }
 
   /**
@@ -64,11 +64,11 @@ function App() {
   }
 
   const checkToken = () => {
-    Logger.log('info', 'Refresh token check');
+    logger.info("Refresh token check");
 
     // Ignore if already refreshing token.
     if (refreshingToken) {
-      Logger.log('info', 'Already refreshing token.');
+      logger.info("Already refreshing token.");
       return;
     }
 
@@ -83,7 +83,7 @@ function App() {
     );
 
     if (!rToken || !exp || !iat) {
-      Logger.log('warn', 'Refresh token, exp or iat not set.');
+      logger.warn("Refresh token, exp or iat not set.");
       return;
     }
 
@@ -94,13 +94,13 @@ function App() {
     // If more than half the time till expire has been passed refresh the token.
     if (now > iat + timeDiff / 2) {
       setRefreshingToken(true);
-      Logger.log('info', 'Refreshing token.');
+      logger.info("Refreshing token.");
 
       ConfigLoader.loadConfig().then((config) => {
         fetch(`${config.apiEndpoint}/v2/authentication/token/refresh`, {
-          method: 'POST',
+          method: "POST",
           headers: {
-            'Content-Type': 'application/json',
+            "Content-Type": "application/json",
           },
           body: JSON.stringify({
             refresh_token: rToken,
@@ -108,7 +108,7 @@ function App() {
         })
           .then((response) => response.json())
           .then((data) => {
-            Logger.log('info', 'Token refreshed.');
+            logger.info("Token refreshed.");
 
             const decodedToken = jwtDecode(data.token);
 
@@ -127,15 +127,14 @@ function App() {
             );
           })
           .catch(() => {
-            Logger.log('error', 'Token refresh error.');
+            logger.error("Token refresh error.");
           })
           .finally(() => {
             setRefreshingToken(false);
           });
       });
     } else {
-      Logger.log(
-        'info',
+      logger.info(
         `Half the time until expire has not been reached. Will not refresh. Token will expire at ${new Date(
           exp * 1000
         ).toISOString()}`
@@ -144,7 +143,7 @@ function App() {
   };
 
   const startContent = (localScreenId) => {
-    Logger.log('info', 'Starting content.');
+    logger.info("Starting content.");
 
     if (contentServiceRef.current !== null) {
       return;
@@ -161,7 +160,7 @@ function App() {
     const entrypoint = `/v2/screens/${localScreenId}`;
 
     document.dispatchEvent(
-      new CustomEvent('startDataSync', {
+      new CustomEvent("startDataSync", {
         detail: {
           screenPath: entrypoint,
         },
@@ -178,7 +177,7 @@ function App() {
   };
 
   const checkLogin = () => {
-    Logger.log('info', 'Check login.');
+    logger.info("Check login.");
 
     const localStorageToken = localStorage.getItem(localStorageKeys.API_TOKEN);
     const localScreenId = localStorage.getItem(localStorageKeys.SCREEN_ID);
@@ -188,14 +187,14 @@ function App() {
     } else {
       ConfigLoader.loadConfig().then((config) => {
         fetch(`${config.apiEndpoint}/v2/authentication/screen`, {
-          method: 'POST',
-          mode: 'cors',
-          credentials: 'include',
+          method: "POST",
+          mode: "cors",
+          credentials: "include",
         })
           .then((response) => response.json())
           .then((data) => {
             if (
-              data?.status === 'ready' &&
+              data?.status === "ready" &&
               data?.token &&
               data?.screenId &&
               data?.tenantKey &&
@@ -221,7 +220,7 @@ function App() {
               localStorage.setItem(localStorageKeys.TENANT_ID, data.tenantId);
 
               startContent(data.screenId);
-            } else if (data?.status === 'awaitingBindKey') {
+            } else if (data?.status === "awaitingBindKey") {
               if (data?.bindKey) {
                 setBindKey(data.bindKey);
               }
@@ -251,7 +250,7 @@ function App() {
   };
 
   const reauthenticateHandler = () => {
-    Logger.log('info', 'Reauthenticate.');
+    logger.info("Reauthenticate.");
 
     localStorage.removeItem(localStorageKeys.API_TOKEN);
     localStorage.removeItem(localStorageKeys.API_TOKEN_EXPIRE);
@@ -277,37 +276,47 @@ function App() {
   };
 
   const checkForUpdates = () => {
-    Logger.log('info', 'Checking for new release timestamp.');
+    logger.info("Checking for new release timestamp.");
 
-    ReleaseLoader.loadConfig().then((config) => {
+    ReleaseLoader.loadConfig().then((release) => {
       if (releaseTimestampRef?.current === null) {
-        releaseTimestampRef.current = config.releaseTimestamp;
-      } else if (releaseTimestampRef?.current !== config.releaseTimestamp) {
-        const redirectUrl = new URL(window.location.href);
-        redirectUrl.searchParams.set(
-          'releaseTimestamp',
-          config.releaseTimestamp
-        );
-        redirectUrl.searchParams.set('releaseVersion', config.releaseVersion);
+        releaseTimestampRef.current = release.releaseTimestamp;
+      } else if (releaseTimestampRef?.current !== release.releaseTimestamp) {
+        if (
+          release.releaseTimestamp !== null &&
+          release.releaseVersion !== null
+        ) {
+          const redirectUrl = new URL(window.location.href);
+          redirectUrl.searchParams.set(
+            "releaseTimestamp",
+            release.releaseTimestamp
+          );
+          redirectUrl.searchParams.set(
+            "releaseVersion",
+            release.releaseVersion
+          );
 
-        window.location.replace(redirectUrl);
+          window.location.replace(redirectUrl);
+        } else {
+          logger.info("Release timestamp or version null, not redirecting.");
+        }
       }
     });
   };
 
   const contentEmpty = () => {
-    Logger.log('info', 'Content empty. Displaying fallback.');
+    logger.info("Content empty. Displaying fallback.");
     setDisplayFallback(true);
   };
 
   const contentNotEmpty = () => {
-    Logger.log('info', 'Content not empty. Displaying content.');
+    logger.info("Content not empty. Displaying content.");
     setDisplayFallback(false);
   };
 
   // ctrl/cmd i will log screen out and refresh
   const handleKeyboard = ({ repeat, metaKey, ctrlKey, code }) => {
-    if (!repeat && (metaKey || ctrlKey) && code === 'KeyI') {
+    if (!repeat && (metaKey || ctrlKey) && code === "KeyI") {
       localStorage.clear();
       window.location.reload();
     }
@@ -318,25 +327,34 @@ function App() {
 
     // Make sure have releaseVersion and releaseTimestamp set in url parameters.
     if (
-      !currentUrl.searchParams.has('releaseVersion') ||
-      !currentUrl.searchParams.has('releaseTimestamp')
+      !currentUrl.searchParams.has("releaseVersion") ||
+      !currentUrl.searchParams.has("releaseTimestamp")
     ) {
       ReleaseLoader.loadConfig().then((release) => {
-        currentUrl.searchParams.set(
-          'releaseTimestamp',
-          release.releaseTimestamp
-        );
-        currentUrl.searchParams.set('releaseVersion', release.releaseVersion);
+        if (
+          release.releaseTimestamp !== null &&
+          release.releaseVersion !== null
+        ) {
+          currentUrl.searchParams.set(
+            "releaseTimestamp",
+            release.releaseTimestamp
+          );
+          currentUrl.searchParams.set("releaseVersion", release.releaseVersion);
 
-        window.history.replaceState(null, '', currentUrl);
+          window.history.replaceState(null, "", currentUrl);
+        } else {
+          logger.info(
+            "Release timestamp or version null, not setting query parameters."
+          );
+        }
       });
     }
 
-    document.addEventListener('screen', screenHandler);
-    document.addEventListener('reauthenticate', reauthenticateHandler);
-    document.addEventListener('contentEmpty', contentEmpty);
-    document.addEventListener('contentNotEmpty', contentNotEmpty);
-    document.addEventListener('keypress', handleKeyboard);
+    document.addEventListener("screen", screenHandler);
+    document.addEventListener("reauthenticate", reauthenticateHandler);
+    document.addEventListener("contentEmpty", contentEmpty);
+    document.addEventListener("contentNotEmpty", contentNotEmpty);
+    document.addEventListener("keypress", handleKeyboard);
 
     checkLogin();
 
@@ -351,13 +369,13 @@ function App() {
     });
 
     return function cleanup() {
-      Logger.log('info', 'Unmounting App.');
+      logger.info("Unmounting App.");
 
-      document.removeEventListener('keypress', handleKeyboard);
-      document.removeEventListener('screen', screenHandler);
-      document.removeEventListener('reauthenticate', reauthenticateHandler);
-      document.removeEventListener('contentEmpty', contentEmpty);
-      document.removeEventListener('contentNotEmpty', contentNotEmpty);
+      document.removeEventListener("keypress", handleKeyboard);
+      document.removeEventListener("screen", screenHandler);
+      document.removeEventListener("reauthenticate", reauthenticateHandler);
+      document.removeEventListener("contentEmpty", contentEmpty);
+      document.removeEventListener("contentNotEmpty", contentNotEmpty);
 
       if (timeoutRef?.current) {
         clearTimeout(timeoutRef.current);
@@ -377,10 +395,10 @@ function App() {
     // Append screenId to current url for easier debugging. If errors are logged in the API's standard http log this
     // makes it easy to see what screen client has made the http call by putting the screen id in the referer http
     // header.
-    if (screen && screen['@id']) {
+    if (screen && screen["@id"]) {
       const url = new URL(window.location.href);
-      url.searchParams.set('screenId', idFromPath(screen['@id']));
-      window.history.replaceState(null, '', url);
+      url.searchParams.set("screenId", idFromPath(screen["@id"]));
+      window.history.replaceState(null, "", url);
     }
 
     ConfigLoader.loadConfig().then((config) => {
@@ -396,7 +414,7 @@ function App() {
         fetch(`${config.apiEndpoint}/v2/tenants/${tenantId}`, {
           headers: {
             authorization: `Bearer ${token}`,
-            'Authorization-Tenant-Key': tenantKey,
+            "Authorization-Tenant-Key": tenantKey,
           },
         })
           .then((response) => response.json())