+++ src/logger/logger.js
@@ -1,22 +1,5 @@
-const winston = require('winston');
+import pino from 'pino';
 
-// @TODO: Add configuration of the logger.
+const logger = pino();
 
-const logger = winston.createLogger({
-  level: 'info',
-  format: winston.format.json(),
-  defaultMeta: { service: 'user-service' },
-  transports: [],
-});
-
-// @TODO: Add based on configuration instead of always.
-logger.add(
-  new winston.transports.Console({
-    format: winston.format.combine(
-      winston.format.colorize(),
-      winston.format.simple()
-    ),
-  })
-);
-
-module.exports = logger;
+export default logger;